% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{OlympicsMedal}
\alias{OlympicsMedal}
\title{Olympics Medal}
\format{
A data frame with several rows and columns:
\describe{
  \item{Country}{Country name}
  \item{Number_of_Athletes}{Total athletes representing each country}
  \item{Gold}{Number of gold medals}
  \item{Silver}{Number of silver medals}
  \item{Bronze}{Number of bronze medals}
  \item{Total}{Total medals won}
  \item{Ratio_Athletes_Per_Medal}{Ratio of athletes per medal}
}
}
\usage{
OlympicsMedal
}
\description{
A dataset containing information on Olympic athletes and medals.
}
\keyword{datasets}
