% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medal_summary.R
\name{summarize_medals}
\alias{summarize_medals}
\title{Summarize Medal Counts}
\usage{
summarize_medals(df)
}
\arguments{
\item{df}{A data frame of medal data.}
}
\value{
A summary table of medals by country.
}
\description{
This function summarizes the medal counts by country.
}
